+++ src/language-html/utils.js
@@ -1,5 +1,52 @@
 "use strict";
 
+// https://html.spec.whatwg.org/multipage/indices.html#attributes-3
+const BOOLEAN_ATTRIBUTES = [
+  "allowfullscreen",
+  "allowpaymentrequest",
+  "async",
+  "autofocus",
+  "autoplay",
+  "checked",
+  "controls",
+  "default",
+  "defer",
+  "disabled",
+  "formnovalidate",
+  "hidden",
+  "ismap",
+  "itemscope",
+  "loop",
+  "multiple",
+  "muted",
+  "nomodule",
+  "novalidate",
+  "open",
+  "readonly",
+  "required",
+  "reversed",
+  "selected",
+  "typemustmatch"
+];
+
+// http://w3c.github.io/html/single-page.html#void-elements
+const VOID_TAGS = [
+  "area",
+  "base",
+  "br",
+  "col",
+  "embed",
+  "hr",
+  "img",
+  "input",
+  "link",
+  "meta",
+  "param",
+  "source",
+  "track",
+  "wbr"
+];
+
 function hasPrettierIgnore(path) {
   const node = path.getValue();
 
@@ -41,6 +88,33 @@ function isWhitespaceOnlyText(node) {
   return node.type === "text" && node.data.trim().length === 0;
 }
 
+function isBooleanAttributeNode(node) {
+  return (
+    node.type === "attribute" && BOOLEAN_ATTRIBUTES.indexOf(node.key) !== -1
+  );
+}
+
+function isVoidTagNode(node) {
+  return node.type === "tag" && VOID_TAGS.indexOf(node.name) !== -1;
+}
+
+function isPreTagNode(node) {
+  return node.type === "tag" && node.name === "pre";
+}
+
+function isTextAreaTagNode(node) {
+  return node.type === "tag" && node.name === "textarea";
+}
+
+function isScriptTagNode(node) {
+  return node.type === "script" || node.type === "style";
+}
+
 module.exports = {
-  hasPrettierIgnore
+  hasPrettierIgnore,
+  isBooleanAttributeNode,
+  isPreTagNode,
+  isScriptTagNode,
+  isTextAreaTagNode,
+  isVoidTagNode
 };
